ENH: Andor cooling does not have to be set on startup
InitializeCoolerState is a new config param that if set to True will turn on the cooler, but if set to False will maintain whatever cooler state it is was in before connection.

There is also RequireCoolTemp for dev purposes which can be set to False if you want to acquire images without it having reached the desired cool temperature.

These two changes replace the previous UseCooling method which forced the cooler on/off at startup.